requirements.txt
requests==2.26.0
python-decouple==3.5
Pillow==8.4.0
pyTelegramBotAPI==4.2.1
moviepy==1.0.3